+++ Kernel/Modules/AgentTicketLockedView.pm
@@ -389,7 +389,7 @@ sub Run {
 
         Filter     => $Filter,
         Filters    => \%NavBarFilter,
-        FilterLink => $LinkFilter,
+        LinkFilter => $LinkFilter,
 
         TitleName  => Translatable('My Locked Tickets'),
         TitleValue => $Filters{$Filter}->{Name},