+++ Kernel/Modules/AgentTicketSearch.pm
@@ -1150,8 +1150,7 @@ sub Run {
                 Bulk      => 1,
                 Limit     => $Self->{SearchLimit},
 
-                Filter     => $Self->{Filter},
-                FilterLink => $FilterLink,
+                Filter => $Self->{Filter},
 
                 OrderBy      => $Self->{OrderBy},
                 SortBy       => $Self->{SortBy},